requirements.txt
pytorch
numpy
scipy
matplotlib
pandas
ipykernel
tqdm
autopep8